requirements.txt
six==1.11.0
websocket-client==0.44.0
